tools/converter_gui/requirements.txt
ultralytics>=8.2.0
onnx>=1.16.0
pyyaml>=6.0.1
